Playtest passthrough - Lower party kill requirements - Fixed a null pointer with purification in shop - Allow spectators to fly in all settings - Fixed fencing sword and reforges - Improve harvest fields debug capabilities - Some bugfixes and balancing

+++ src/me/neoblade298/neorogue/equipment/armor/ClothBindings.java
@@ -12,7 +12,7 @@ import me.neoblade298.neorogue.session.fight.trigger.TriggerResult;
 
 public class ClothBindings extends Equipment {
 	private static final String ID = "ClothBindings";
-	private double health;
+	private int health;
 	
 	public ClothBindings(boolean isUpgraded) {
 		super(ID, "Cloth Bindings", isUpgraded, Rarity.COMMON, EquipmentClass.WARRIOR,
